+++ src/main/java/olter/balls/campaigns/campaign/model/CampaignRepository.java
@@ -2,4 +2,4 @@ package olter.balls.campaigns.campaign.model;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CampaignRepository extends JpaRepository<CampaignEntity, Integer> {}
+public interface CampaignRepository extends JpaRepository<CampaignEntity, Long> {}
