+++ src/main/java/olter/balls/database/ancestries/ancestry/model/AncestryRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AncestryRepository extends JpaRepository<AncestryEntity, Integer> {
+public interface AncestryRepository extends JpaRepository<AncestryEntity, Long> {
   Optional<AncestryEntity> findByName(String name);
 }
