+++ src/main/java/olter/balls/database/books/model/BookRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BookRepository extends JpaRepository<BookEntity, Integer> {
+public interface BookRepository extends JpaRepository<BookEntity, Long> {
   Optional<BookEntity> findByShortName(String shortName);
 }
